+++ lib/middleman-akcms/manipulator.rb
@@ -2,9 +2,10 @@
 module Middleman::Akcms
   ## base class of manipulators
   class Manipulator
-    attr_reader :controller
+    attr_reader :controller, :sitemap
     def initialize(controller)
       @controller = controller
+      @sitemap = controller.extension.app.sitemap
     end
     def manipulate_resource_list(resources)
       resources